regression.js and splice
Adds splice to core (to fix inverse which calls it and says the function
does not exist, do you have another way of implementing that?) and gives
the regression js. Notice that regress and regresst are not functional
since there is some error dealing with the inverse. Matrixmult and
matrixsubtract were included because I was having issues with the
supplied mult and subtract when giving them matrix arguments.

I will not have time for a long while so I putting this all up for
someone to complete or to come back to later.