Add tests for various linear algebra functions
Additionally, fixed a small bug that prevented .dot() and .angle() from
being called on a jStat object, as opposed to as a static function.

There are still untested functions in the linear algebra library, and it
does not make sense that all of the linear algebra functions can be
called on a jStat() object but only accept plain matricies as arguments.
This future change will be made easier with fuller test coverage.

Reviewed-by: Trevor Norris <trev.norris@gmail.com>