Fix #106
Instead of moving to Grunt or Gulp, put the scripts we need into our
package.json file. To maintain backwards compatibility with existing
build setups, the Makefile is retained, but now refers to scripts
defined in package.json using calls to npm run-script.

This removes the dependence on make for users on windows or platforms
without a current or up to date version of make, and should make
contributing to the project easier.

I left the defaults for vows (--isolate) that were already in
package.json, even though those were different from the defaults in the
Makefile. Perosnally, I prefer the less verbose vows output, and would
love to change this.