Add a version number, close #105
I tried to predict which version this would land in, but of course it
will need to be bumped every release.

This would fulfill #105 but does add another step to the release
process. Personally, I am indifferent to whether or not the library
should provide a .version(), but if not, we should explicitly close
issue #105.