Update documentation and add test for mode(true)
The documentation previously specified that jStat(<two dimensional
array>).mode(true) returned 'the mode of the array', i.e. the most
common row. However, inspection of the .mode() function shows that
.mode(true) instead treats the two-dimensional array as one dimensional.

I believe this is a reasonable method option; as far as I can tell there
is no commonly accepted definition for 'the mode of an array'. If there
is strong demand for 'finding the most common data point in my two
dimensional data', it would be better to define it explicitly.

Add a test to verify this behavior and update the documentation to
reflect this.

Fixes: https://github.com/jstat/jstat/issues/50
PR-URL: https://github.com/jstat/jstat/pull/144
Reviewed-by: Trevor Norris <trev.norris@gmail.com>