Add median function for beta distribution
The median function was not yet implemented for the beta distribution.
This adds one using a straightforward implementation based on the
already existing quantile function.

Adding testing for beta.median and fixing testing for beta.mode

Adding tests for beta.median, comparing it to qbeta() in R. Also fixing
the tests for beta.mode which had a significant digits problem and a
typo.

PR-URL: https://github.com/jstat/jstat/pull/151
Reviewed-By: James Gibson <james.cameron.gibson@gmail.com>