Fix issue #97
When x = 0 and df1 <= 2, the pdf was still returning NaN. There were a
couple of checks in the code that appeared to be intended to deal with
this situation, but they weren't conditional on x = 0, when they should,
and they were conditional on the value of df2, when they shouldn't. This
fixes those problems.