Add quantile function for triangular distribution
The triangular distribution did not have a quantile function (inverse
cumulative distribution function) defined. The math is based on
definition 25 on http://www.math.uah.edu/stat/special/Triangle.html, but
it has been converted that definition to be in terms of the a, b, c
parameterization instead of the a, w, p parameterization.

PR-URL: https://github.com/jstat/jstat/pull/149
Reviewed-by: Trevor Norris <trev.norris@gmail.com>