Add Browser Usage to README
Users requested that the use of jStat in the browser be clarified. This
information was (sort of) in the README, but it could be improved. So,
we re-order the using & building sections, as there are more library users
than library developments. A note clarifying the usage of jStat in the
browser is added, and a note addressing what jStat is most useful for.

Finally, add a note about installing from NPM & an NPM badge, since NPM seems
to have captured the most mindshare of the JS dependency managers.

Fixes: https://github.com/jstat/jstat/issues/184
PR-URL: https://github.com/jstat/jstat/pull/187
Reviewed-By: Trevor Norris <trev.norris@gmail.com>