Fix issue #219 - Missing calculation of adjusted R^2
The Ordinary Least Squares module was not computing the adjusted R^2
correctly.

Resolve this issue, additionally adding a test for the correct
calculation, and further document the test, code, and generated
documentation to clarify for users which adjusted R^2 value is being
calcualted.

Thank you to EloquentData for reporting the issue.

Fixes: https://github.com/jstat/jstat/issues/219
PR-URL: https://github.com/jamescgibson/jstat/pull/221